move redundant F(x) to helper-func.js, annotation/cleanup in prep for refactoring for API
